split GC code into a separate C file
