+++ c/main.c
@@ -44,6 +44,7 @@
 #include "foo.h"
 #include "system.h"
 #include "ext.h"
+#include "mark-and-sweep.h"
 
 size_t min_size(size_t a, size_t b) {
   if (a <= b) {
@@ -56,21 +57,6 @@ size_t min_size(size_t a, size_t b) {
 #define PTR(type, datum) \
   ((struct type*)datum.ptr)
 
-#if 0
-# define FOO_DEBUG(...) { fprintf(stderr, __VA_ARGS__); fprintf(stderr, "\n"); fflush(stderr); }
-#else
-# define FOO_DEBUG(...)
-#endif
-
-#define FOO_XXX(...) { fprintf(stderr, "XXX: "); fprintf(stderr, __VA_ARGS__); fprintf(stderr, "\n"); fflush(stderr); }
-
-struct FooContext;
-void* foo_alloc(struct FooContext*, size_t bytes);
-
-struct Foo foo_panic(struct FooContext* ctx, struct Foo message) __attribute__((noreturn));
-struct Foo foo_panicf(struct FooContext* ctx, const char* fmt, ...) __attribute__((noreturn));
-
-void foo_unimplemented(const char* message) __attribute__ ((noreturn));
 void foo_unimplemented(const char* message) {
   printf("UNIMPLEMENTED: %s", message);
   fflush(stdout);
@@ -78,40 +64,12 @@ void foo_unimplemented(const char* message) {
   _Exit(1);
 }
 
-void foo_abort(const char* message) __attribute__ ((noreturn));
 void foo_abort(const char* message) {
   perror(message);
   fflush(stdout);
   fflush(stderr);
   _Exit(1);
 }
-
-union FooDatum {
-  void* ptr;
-  int64_t int64;
-  double float64;
-  int64_t boolean; // Make sure there's no junk in high bits
-};
-
-struct Foo {
-  struct FooClass* class;
-  union FooDatum datum;
-};
-
-enum FooAllocType {
-  STATIC,
-  HEAP
-};
-
-struct FooHeader {
-  enum FooAllocType allocation;
-};
-
-struct FooBytes {
-  struct FooHeader header;
-  size_t size;
-  uint8_t data[];
-};
 
 #define FOO_CSTRING(literal) \
   ((struct FooBytes*) \
@@ -158,69 +116,6 @@ struct FooSelector* foo_intern(struct FooBytes* name) {
   return foo_intern_new_selector(name);
 }
 
-struct FooArray {
-  struct FooHeader header;
-  size_t size;
-  struct Foo data[];
-};
-
-// FIXME: Don't like defining this in C.
-struct FooFile {
-  struct FooHeader header;
-  struct FooBytes* pathname;
-  size_t mode;
-};
-
-// FIXME: Don't like defining this in C.
-struct FooFileStream {
-  struct FooHeader header;
-  struct FooBytes* pathname;
-  FILE* ptr;
-};
-
-struct FooCleanup;
-typedef void (*FooCleanupFunction)(struct FooContext*, struct FooCleanup*);
-
-struct FooCleanup {
-  FooCleanupFunction function;
-  struct FooCleanup* next;
-};
-
-struct FooFinally {
-  struct FooCleanup cleanup;
-  struct FooClosure* closure;
-};
-
-struct FooUnbind {
-  struct FooCleanup cleanup;
-  size_t index;
-  struct Foo value;
-};
-
-enum FooContextType {
-    METHOD_CONTEXT,
-    CLOSURE_CONTEXT,
-    UNWIND_CONTEXT,
-    ROOT_CONTEXT
-};
-
-struct FooContext {
-  enum FooContextType type;
-  uint32_t depth;
-  const struct FooMethod* method;
-  struct Foo receiver;
-  struct FooContext* sender;
-  struct FooContext* outer_context;
-  // FIXME: Doesn't really belong in context, but easier right now.
-  struct FooArray* vars;
-  struct FooCleanup* cleanup;
-  struct Foo return_value;
-  // Only for methods, for others this is NULL.
-  jmp_buf* ret;
-  size_t size;
-  struct Foo frame[];
-};
-
 struct FooContext* foo_alloc_context(struct FooContext* sender, size_t size) {
   struct FooContext* ctx
     = foo_alloc(sender, sizeof(struct FooContext) + size * sizeof(struct Foo));
@@ -236,9 +131,6 @@ char* foo_debug_context(struct FooContext* ctx) {
            ctx->type, ctx->size, ctx->outer_context);
   return s;
 }
-
-typedef struct Foo (*FooMethodFunction)(const struct FooMethod*, struct FooContext*);
-typedef struct Foo (*FooClosureFunction)(struct FooContext*);
 
 struct Foo foo_lexical_ref(struct FooContext* context, size_t index, size_t frameOffset) {
   struct FooContext* context0 = context;
@@ -269,24 +161,6 @@ struct Foo foo_lexical_set(struct FooContext* context, size_t index, size_t fram
   context->frame[index] = value;
   return value;
 }
-
-struct FooMethod {
-  struct FooClass* class; // FIXME: rename to home
-  struct FooSelector* selector;
-  size_t argCount;
-  size_t frameSize;
-  // Native method functions directly implement the method
-  // Object method functions send #invoke:inContext: to the object
-  FooMethodFunction function;
-  struct Foo object;
-};
-
-struct FooClosure {
-  struct FooContext* context;
-  size_t argCount;
-  size_t frameSize;
-  FooClosureFunction function;
-};
 
 // Forward declarations for classs are in generated_classes, but we're going
 // to define a few builtin ctors first that need some of them.
@@ -386,17 +260,6 @@ void foo_unbind(struct FooContext* sender, struct FooCleanup* cleanup) {
   sender->vars->data[unbind->index] = unbind->value;
 }
 
-typedef void (*FooMarkFunction)(void* ptr);
-void foo_mark_array(void* ptr);
-void foo_mark_class(void* ptr);
-void foo_mark_none(void* ptr);
-
-struct FooClassList {
-  struct FooHeader header;
-  size_t size;
-  void* data[];
-};
-
 struct FooClassList* FooClassList_alloc(struct FooContext* sender, size_t size) {
   struct FooClassList* list
     = foo_alloc(sender, sizeof(struct FooClassList) + size * sizeof(void*));
@@ -404,12 +267,6 @@ struct FooClassList* FooClassList_alloc(struct FooContext* sender, size_t size) 
   list->size = size;
   return list;
 }
-
-struct FooLayout {
-  struct FooHeader header;
-  FooMarkFunction mark;
-  size_t size;
-};
 
 struct FooLayout TheEmptyLayout = {
   .header = { .allocation = STATIC },
@@ -438,17 +295,6 @@ struct FooLayout* foo_Layout_new(struct FooContext* sender, size_t size) {
   layout->size = size;
   return layout;
 }
-
-struct FooClass {
-  struct FooHeader header;
-  struct FooBytes* name;
-  struct FooClass* metaclass;
-  struct FooClassList* inherited;
-  struct FooLayout* layout;
-  FooMarkFunction mark;
-  size_t size;
-  struct FooMethod methods[];
-};
 
 struct Foo foo_class_new(struct FooContext* ctx) {
   struct FooClass* theClass = PTR(FooClass, ctx->frame[0].datum);
@@ -1014,331 +860,6 @@ void fooinit(void) {
   _setmode(_fileno(stdout), O_BINARY);
   _setmode(_fileno(stderr), O_BINARY);
 #endif
-}
-
-/**
-   GC
-
- */
-
-bool trace_gc = false;
-size_t gc_trace_depth = 0;
-#define ENTER_TRACE(...) if (trace_gc) { fprintf(stderr, "\n"); for(size_t i = 0; i < gc_trace_depth; i++) fprintf(stderr, "  "); fprintf(stderr, "%zu: ", gc_trace_depth); fprintf(stderr, __VA_ARGS__); gc_trace_depth++; }
-#define EXIT_TRACE() if (trace_gc) { gc_trace_depth--; if (!gc_trace_depth) fprintf(stderr, "\n"); }
-
-#if 0
-#define DEBUG_GC(...) { fprintf(stderr, __VA_ARGS__); fflush(stderr); }
-#else
-#define DEBUG_GC(...)
-#endif
-
-enum FooMark {
-  DEAD = 0,
-  LIVE = 1,
-};
-
-struct FooAlloc {
-  enum FooMark mark;
-  struct FooAlloc* next;
-  size_t size;
-  char data[];
-};
-
-bool foo_mark_live(void* ptr) {
-  ENTER_TRACE("mark_live %p", ptr);
-  const size_t offset = offsetof(struct FooAlloc, data);
-  struct FooAlloc* alloc = (void*)((char*)ptr-offset);
-  bool new_mark = alloc->mark == DEAD;
-  alloc->mark = LIVE;
-  EXIT_TRACE();
-  return new_mark;
-}
-
-void foo_mark_ptr(void* ptr) {
-  ENTER_TRACE("mark_ptr");
-  foo_mark_live(ptr);
-  EXIT_TRACE();
-}
-
-void foo_mark_bytes(void* ptr) {
-  ENTER_TRACE("mark_bytes");
-  struct FooBytes* bytes = ptr;
-  if (bytes->header.allocation == HEAP) {
-    foo_mark_live(ptr);
-  }
-  EXIT_TRACE();
-}
-
-void foo_mark_file(void* ptr) {
-  ENTER_TRACE("mark_file");
-  struct FooFile* file = ptr;
-  if (file->header.allocation == HEAP && foo_mark_live(file)) {
-    foo_mark_bytes(file->pathname);
-  }
-  EXIT_TRACE();
-}
-
-void foo_mark_filestream(void* ptr) {
-  ENTER_TRACE("mark_bytes");
-  struct FooFileStream* stream = ptr;
-  if (stream->header.allocation == HEAP && foo_mark_live(stream))  {
-    foo_mark_bytes(stream->pathname);
-  }
-  EXIT_TRACE();
-}
-
-void foo_mark_context(struct FooContext* ctx);
-
-void foo_mark_object(struct Foo obj) {
-  ENTER_TRACE("mark_object");
-  if (obj.class) {
-    DEBUG_GC(" %p (%s)", obj.datum.ptr, obj.class->name->data);
-    foo_mark_class(obj.class);
-    obj.class->mark(obj.datum.ptr);
-  } else {
-    assert(!obj.datum.int64);
-  }
-  EXIT_TRACE();
-}
-
-void foo_mark_none(void* ptr) {
-  (void)ptr;
-}
-
-void foo_mark_oops(void* ptr) {
-  foo_abort("Oops");
-}
-
-void foo_mark_array(void* ptr) {
-  ENTER_TRACE("mark_array");
-  struct FooArray* array = ptr;
-  if (array->header.allocation == HEAP && foo_mark_live(array)) {
-    for (size_t i = 0; i < array->size; i++) {
-      foo_mark_object(array->data[i]);
-    }
-  }
-  EXIT_TRACE();
-}
-
-void foo_mark_instance(void* ptr) {
-  ENTER_TRACE("mark_instance");
-  struct FooArray* array = ptr;
-  if (array->header.allocation == HEAP && foo_mark_live(array)) {
-    for (size_t i = 0; i < array->size; i++) {
-      foo_mark_object(array->data[i]);
-    }
-  }
-  EXIT_TRACE();
-}
-
-void foo_mark_layout(void* ptr) {
-  struct FooLayout* layout = ptr;
-  bool is_empty = layout == &TheEmptyLayout;
-  bool is_class = layout == &TheClassLayout;
-  (void)is_empty;
-  (void)is_class;
-  ENTER_TRACE("mark_layout (%s)",
-              is_empty ? "empty" : (is_class ? "class" : "object"));
-  if (layout->header.allocation == HEAP) {
-    foo_mark_live(layout);
-  }
-  EXIT_TRACE();
-}
-
-void foo_mark_class_list(void* ptr) {
-  struct FooClassList* list = ptr;
-  ENTER_TRACE("mark_class_list %p (size=%zu)", list, list->size);
-  if (list->header.allocation == HEAP && foo_mark_live(list)) {
-    for (size_t i = 0; i < list->size; i++) {
-      foo_mark_class(list->data[i]);
-    }
-  }
-  EXIT_TRACE();
-}
-
-void foo_mark_class(void* ptr)
-{
-  struct FooClass* class = ptr;
-  ENTER_TRACE("mark_class %p (%s)", class, class->name->data);
-  assert(class);
-  if (class->header.allocation == HEAP && foo_mark_live(class)) {
-    foo_mark_bytes(class->name);
-    foo_mark_class(class->metaclass);
-    foo_mark_class_list(class->inherited);
-    foo_mark_layout(class->layout);
-    for (size_t i = 0; i < class->inherited->size; i++) {
-      struct FooClass* other = class->inherited->data[i];
-      if (other)
-        foo_mark_class(other);
-    }
-    for (size_t i = 0; i < class->size; i++) {
-      struct FooMethod* method = &class->methods[i];
-      if (method->object.class)
-        foo_mark_object(method->object);
-    }
-  }
-  EXIT_TRACE();
-}
-
-void foo_mark_closure(void* ptr) {
-  ENTER_TRACE("mark_closure");
-  struct FooClosure* closure = ptr;
-  if (foo_mark_live(closure)) {
-    foo_mark_context(closure->context);
-  }
-  EXIT_TRACE();
-}
-
-void foo_mark_cleanup(struct FooCleanup* cleanup) {
-  ENTER_TRACE("mark_cleanup");
-  if (!cleanup) {
-    goto exit;
-  }
-  if (cleanup->function == foo_finally) {
-    foo_mark_closure(((struct FooFinally*)cleanup)->closure);
-    goto exit;
-  }
-  if (cleanup->function == foo_unbind) {
-    foo_mark_object(((struct FooUnbind*)cleanup)->value);
-    goto exit;
-  }
- exit:
-  EXIT_TRACE();
-}
-
-void foo_mark_context(struct FooContext* ctx) {
-  ENTER_TRACE("mark_context");
-  if (ctx) {
-    DEBUG_GC(" depth: %u, size: %zu", ctx->depth, ctx->size);
-    if (ctx->type == METHOD_CONTEXT) {
-      DEBUG_GC(" selector: %s", ctx->method->selector->name->data);
-    }
-  }
-  if (ctx && foo_mark_live(ctx)) {
-    foo_mark_object(ctx->receiver);
-    foo_mark_context(ctx->sender);
-    foo_mark_context(ctx->outer_context);
-    // vars are shared between all contexts, foo_mark() processes them directly.
-    foo_mark_cleanup(ctx->cleanup);
-    foo_mark_object(ctx->return_value);
-    for (size_t i = 0; i < ctx->size; i++) {
-      // printf("\n[%zu]", i);
-      foo_mark_object(ctx->frame[i]);
-    }
-  }
-  EXIT_TRACE();
-}
-
-struct FooAlloc* allocations = NULL;
-static size_t allocation_count_since_gc = 0;
-static size_t allocation_bytes_since_gc = 0;
-static size_t allocation_bytes = 0;
-static size_t allocation_count = 0;
-
-// For testing: low threshold so that GC gets exercised way more.
-// const size_t gc_threshold = 1024;
-const size_t gc_threshold = 1024 * 1024 * 64;
-bool gc_verbose = false;
-
-void foo_sweep() {
-  struct FooAlloc* head = allocations;
-  size_t freed_count = 0, freed_bytes = 0;
-  size_t live_count = 0, live_bytes = 0;
-
-  size_t n = 0;
-  struct FooAlloc* prev = NULL;
-  while (head) {
-    n++;
-    struct FooAlloc* next = head->next;
-    if (head->mark == DEAD) {
-      if (!prev) {
-        allocations = next;
-      } else {
-        prev->next = next;
-      }
-      freed_bytes += head->size;
-      freed_count += 1;
-      free(head);
-    } else {
-      prev = head;
-      live_count += 1;
-      live_bytes += head->size;
-    }
-    head = next;
-  }
-
-  assert(allocation_count == n);
-
-  if (freed_count > 0) {
-    allocation_bytes -= freed_bytes;
-    allocation_count -= freed_count;
-
-    if (gc_verbose) {
-      double mb = 1024.0 * 1024.0;
-      fprintf(stderr,
-              "-- %.2fMB in %zu objects allocated since last gc\n"
-              "-- %.2fMB in %zu objects collected\n"
-              "-- %.2fMB in %zu objects remain\n",
-              allocation_bytes_since_gc / mb, allocation_count_since_gc,
-              freed_bytes / mb, freed_count,
-              allocation_bytes / mb, allocation_count);
-      fflush(stderr);
-    }
-
-    assert(live_count == allocation_count);
-    assert(live_bytes == allocation_bytes);
-
-    allocation_count_since_gc = 0;
-    allocation_bytes_since_gc = 0;
-  }
-}
-
-void foo_gc(struct FooContext* ctx) {
-  FOO_DEBUG("/foo_gc begin");
-  ENTER_TRACE("GC\n");
-
-  // Mark everything dead
-  struct FooAlloc* head = allocations;
-  while (head) {
-    head->mark = DEAD;
-    head = head->next;
-  }
-
-  // Mark everything from ctx live
-  if (ctx->vars) {
-    ENTER_TRACE("vars");
-    foo_mark_array(ctx->vars);
-    EXIT_TRACE();
-  }
-  foo_mark_context(ctx);
-
-  // Free dead things
-  foo_sweep();
-
-  EXIT_TRACE();
-  FOO_DEBUG("/foo_gc end");
-}
-
-void* foo_alloc(struct FooContext* sender, size_t size) {
-  if (allocation_bytes_since_gc > gc_threshold && sender) {
-    foo_gc(sender);
-  }
-  size_t bytes = sizeof(struct FooAlloc) + size;
-  struct FooAlloc* p = calloc(1, bytes);
-  if (!p) {
-    foo_abort("calloc");
-  }
-  p->next = allocations;
-  p->size = bytes;
-  p->mark = LIVE;
-  allocations = p;
-
-  allocation_bytes_since_gc += bytes;
-  allocation_bytes += bytes;
-  allocation_count_since_gc += 1;
-  allocation_count += 1;
-
-  return p->data;
 }
 
 const size_t FooFile_READ      = 0b0001;
